route_planner/requirements.txt
absl-py==0.7.0
actionlib==1.11.13
angles==1.9.11
appnope==0.1.0
astetik==1.9.8
astor==0.7.1
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bondpy==1.8.3
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
catkin==0.7.14
catkin-pkg==0.4.10
certifi==2019.3.9
chances==0.1.6
chardet==3.0.4
cv-bridge==1.12.8
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
docutils==0.14
dynamic-reconfigure==1.5.50
em==0.4.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.2
gazebo-plugins==2.5.18
gazebo-ros==2.5.18
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.7
geonamescache==1.0.1
grpcio==1.19.0
h5py==2.9.0
idna==2.8
image-geometry==1.12.8
interactive-markers==1.11.4
ipython==5.8.0
ipython-genutils==0.2.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kerasplotlib==0.1.4
kiwisolver==1.0.1
laser-geometry==1.6.4
Markdown==3.0.1
matplotlib==2.2.4
message-filters==1.12.14
mock==2.0.0
numpy==1.16.2
pandas==0.24.1
pathlib2==2.3.3
patsy==0.5.1
pbr==5.1.3
pexpect==4.6.0
phased-lstm-keras==1.0.2
pickleshare==0.7.5
pkg-resources==0.0.0
pluginlib==1.11.3
progressbar==2.5
prompt-toolkit==1.0.15
protobuf==3.7.0
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.1
pyquaternion==0.9.5
python-dateutil==2.8.0
python-qt-binding==0.3.4
pytz==2018.9
PyYAML==3.13
qt-dotgraph==0.3.11
qt-gui==0.3.11
qt-gui-cpp==0.3.11
qt-gui-py-common==0.3.11
requests==2.21.0
resource-retriever==1.12.4
rosbag==1.12.14
rosboost-cfg==1.14.4
rosclean==1.14.4
roscreate==1.14.4
rosgraph==1.12.14
roslaunch==1.12.14
roslib==1.14.4
roslint==0.11.0
roslz4==1.12.14
rosmake==1.14.4
rosmaster==1.12.14
rosmsg==1.12.14
rosnode==1.12.14
rosparam==1.12.14
rospkg==1.1.7
rospy==1.12.14
rosservice==1.12.14
rostest==1.12.14
rostopic==1.12.14
rosunit==1.14.4
roswtf==1.12.14
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.9
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.13
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.4.10
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rviz==1.12.16
scandir==1.10.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
sensor-msgs==1.12.7
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
statsmodels==0.9.0
subprocess32==3.5.3
talos==0.4.8
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.20
tf2-kdl==0.5.20
tf2-py==0.5.20
tf2-ros==0.5.20
topic-tools==1.12.14
tqdm==4.31.1
traitlets==4.3.2
transformations==2019.2.20
ujson==1.35
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
wrangle==0.6.2
xacro==1.11.3
